Graph serialization
Add support for serialization of graph to/from proto and to/from file.
