+++ temporian/core/serialize.py
@@ -1,0 +1,30 @@
+# Copyright 2021 Google LLC.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Serialization / unserialization of a processor."""
+
+from typing import List, Set
+
+from temporian.core.data.event import Event
+
+from temporian.proto import core_pb2 as pb
+from temporian.core import processor
+
+
+def serialize(src: processor.Preprocessor) -> pb.Processor:
+    pass
+
+
+def unserialize(src: pb.Processor) -> processor.Preprocessor:
+    pass
