+++ temporian/core/test/BUILD
@@ -3,6 +3,18 @@ package(
     licenses = ["notice"],
 )
 
+py_library(
+    name = "utils",
+    srcs = ["utils.py"],
+    srcs_version = "PY3",
+    deps = [
+        "//temporian/core/data:event",
+        "//temporian/core/data:feature",
+        "//temporian/core/data:sampling",
+        "//temporian/core/operators:base",
+    ],
+)
+
 # Tests
 # =====
 
@@ -26,7 +38,27 @@ py_test(
     deps = [
         # absl/logging dep,
         # absl/testing:absltest dep,
+        ":utils",
         "//temporian/core:processor",
+        "//temporian/core/data:dtype",
+        "//temporian/core/data:event",
+        "//temporian/core/data:feature",
+        "//temporian/core/data:sampling",
+        "//temporian/core/operators:base",
+        "//temporian/proto:core_py_proto",
+    ],
+)
+
+py_test(
+    name = "serialize_test",
+    srcs = ["serialize_test.py"],
+    srcs_version = "PY3",
+    deps = [
+        # absl/logging dep,
+        # absl/testing:absltest dep,
+        ":utils",
+        "//temporian/core:processor",
+        "//temporian/core:serialize",
         "//temporian/core/data:dtype",
         "//temporian/core/data:event",
         "//temporian/core/data:feature",
